pyproject.toml
[build-system]
# Minimum requirements for the build system to execute.
requires = [
    "setuptools",
    "wheel"]  # PEP 508 specifications.

[project.optional-dependencies]
test = [
    'pytest>=6',
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = [
    "tests"
]

[project]
name = "CoregisterUtility"
authors = [
    {name = "Nick Anthony", email = "nickmanthony@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.7"
dynamic = ["version", "description"]

dependencies = [
    "pyimagej",
    "numpy",
    "pyqt >= 5",
    "scikit-image >= 0.19"
]
